Const correctness fix. Warning fixes.

+++ PolyEngine/Engine/Src/InputSystem.cpp
@@ -34,7 +34,7 @@ void InputSystem::InputPhase(World* world)
 			break;
 		case eInputEventType::WHEELMOVE:
 			//gConsole.LogDebug("MouseWheelMove: {}", ev.Pos.X);
-			com.CurrWheel += ev.Pos.X;
+			com.CurrWheel += (int)ev.Pos.X;
 			break;
 		case eInputEventType::_COUNT:
 			HEAVY_ASSERTE(false, "_COUNT enum value passed to InputEventQueue::Push(), which is an invalid value");
